restructured code into better module format
