+++ classroom_voter/tests/test_polltypes.py
