+++ classroom_voter_harrismcc.egg-info/requires.txt
@@ -1,0 +1,2 @@
+pycryptodome
+pycryptodomex
