+++ setup.py
@@ -4,8 +4,8 @@ with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="classroom_voter_harrismcc", # Replace with your own username
-    version="0.0.1",
+    name="classroom-voter-harrismcc", # Replace with your own username
+    version="0.0.2",
     author="Example Author",
     author_email="author@example.com",
     description="A small example package",
@@ -19,4 +19,8 @@ setuptools.setup(
         "Operating System :: OS Independent",
     ],
     python_requires='>=3.6',
+    install_requires=[
+        'pycryptodome',
+        'pycryptodomex'
+    ]   
 )
